move options into spec-directory folder
